split app creation and app launch
